ContactForm component created
