+++ frontend_react/index.html
@@ -50,12 +50,13 @@
   <link rel="icon" sizes="32x32" type="image/png" href="/src/assets/favicon/favicon-32x32-seochecker-313.png">
   <link rel="icon" sizes="16x16" type="image/png" href="/src/assets/favicon/favicon-16x16-seochecker-313.png">
 
-  <script src="https://www.google.com/recaptcha/api.js" async defer></script>
+
 </head>
 
 <body>
   <div id="root"></div>
   <script type="module" src="/src/main.jsx"></script>
+  <script src="https://www.google.com/recaptcha/api.js" async defer></script>
 </body>
 
 </html>
