+++ frontend_react/package.json
@@ -14,6 +14,7 @@
     "node-sass": "^7.0.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-google-recaptcha": "^2.1.0",
     "react-hook-form": "^7.32.0",
     "react-icons": "^4.3.1",
     "react-tooltip": "^4.2.21"
