+++ frontend_react/src/components/index.js
@@ -1,9 +1,11 @@
 import NavigationDots from './NavigationDots';
 import SocialMedia from './SocialMedia';
 import Navbar from './Navbar/Navbar';
+import ContactForm from './ContactForm/ContactForm';
 
 export {
   NavigationDots,
   SocialMedia,
   Navbar,
+  ContactForm,
 };
