+++ frontend_react/src/container/Footer/Footer.jsx
@@ -1,40 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import { images } from '../../constants';
 import { AppWrap, MotionWrap } from '../../wrapper';
-import { client } from '../../client';
+import { ContactForm } from '../../components';
 import './Footer.scss';
 
 const Footer = () => {
-  const [formData, setFormData] = useState({ name: '', company: '', email: '', message: '' });
-  const [isFormSubmitted, setIsFormSubmitted] = useState(false);
-  const [loading, setLoading] = useState(false);
-
-  const { username, company, email, message } = formData;
-
-  const handleChangeInput = (e) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
-
-  const handleSubmit = () => {
-    setLoading(true);
-
-    const contact = {
-      _type: 'contact',
-      name: formData.username,
-      company: formData.company,
-      email: formData.email,
-      message: formData.message,
-    };
-
-    client.create(contact)
-      .then(() => {
-        setLoading(false);
-        setIsFormSubmitted(true);
-      })
-      .catch((err) => console.log(err));
-  };
 
   return (
     <>
@@ -77,38 +48,7 @@ const Footer = () => {
           <a href="tel:+1 (123) 456-7890" className="p-text">+1 (123) 456-7890</a>
         </div>
       </div>
-      {!isFormSubmitted ? (
-        <form 
-        className="app__footer-form app__flex" name="contact" action="?" method="POST">
-          <div className="app__flex">
-            <input className="p-text" type="text" placeholder="Your Name" name="username" value={username} onChange={handleChangeInput} />
-          </div>
-          <div className="app__flex">
-            <input className="p-text" type="text" placeholder="Your Company" name="company" value={company} onChange={handleChangeInput} />
-          </div>
-          <div className="app__flex">
-            <input className="p-text" type="email" placeholder="Your Email" name="email" value={email} onChange={handleChangeInput} />
-          </div>
-          <div>
-            <textarea
-              className="p-text"
-              placeholder="Your Message"
-              value={message}
-              name="message"
-              onChange={handleChangeInput}
-            />
-          </div>
-          <div className="g-recaptcha" data-sitekey="6LdArWggAAAAAEQI4wfaODgc7ImRzJaZQS2us4gp"></div>
-          <button type="button" className="p-text" onClick={handleSubmit}  
-          data-action='submit'>{!loading ? 'Send Message' : 'Sending...'}</button>
-        </form>
-      ) : (
-        <div>
-          <h3 className="head-text">
-            Thank you for getting in touch!
-          </h3>
-        </div>
-      )}
+      <ContactForm/>
     </>
   );
 };
